Add source for dicom2dump tool.
	* fabien/TODO: Simple TODO reminder.
	* fabien/bin/Makefile: Makefile to compile dicom binaries.
	* fabien/bin/dicom2dump.cc: Binary to convert dicom to dump.
	* fabien/bin/dicom_mask.cc: Binary to extract simple mask
	  from dicom images.
	* fabien/igr/Makefile: Update.
	* fabien/igr/launch2d.sh: Update.
	* fabien/igr/launch3d.sh: Update.
	* fabien/igr/seg2d.cc: Update.
	* fabien/igr/seg3d.cc: Update.
	* fabien/igr/seg_vol_irm.hh: Implement different threshold
	  techniques: double and deviation.

git-svn-id: https://svn.lrde.epita.fr/svn/oln/trunk@3482 4aad255d-cdde-0310-9447-f3009e2ae8c0

+++ milena/sandbox/fabien/igr/launch2d.sh
@@ -3,8 +3,9 @@
 process_file ()
 {
       ./seg2d $1
-      if [ -f result.pbm ]; then
-	mv result.pbm results/${2}_06_result.pbm
+      if [ -f result_double.pbm ]; then
+	mv result_double.pbm results/${2}_06_result_double.pbm
+	mv result_deviation.pbm results/${2}_07_result_deviation.pbm
       fi
 
       if [ -f bg.pbm ]; then
@@ -14,10 +15,15 @@ process_file ()
 	mv bg_closed.pbm results/${2}_02_bg_closed.pbm
 	mv obj_closed.pbm results/${2}_04_obj_closed.pbm
 
-	mv regions_color.ppm results/${2}_05_regions_color.ppm
+	if [ -f regions_color.ppm ]; then
+	  mv regions_color.ppm results/${2}_05_regions_color.ppm
+	fi
 
-	mv bg_histo_norm.plot results/${2}_bg.plot
-	mv obj_histo_norm.plot results/${2}_obj.plot
+	mv histo.plot results/${2}.plot
+	mv bg_histo.plot results/${2}_bg.plot
+	mv obj_histo.plot results/${2}_obj.plot
+	mv bg_histo_norm.plot results/${2}_p_bg.plot
+	mv obj_histo_norm.plot results/${2}_p_obj.plot
       fi
 }
